geolocation(fix): put code on according directories and files
